# Optimize CSV marshal: