- Remove a ton of map lookups
- Pre-allocate arrays

Found in profiling that `MarshalCSV` was 23% of my runtime, with ~9% in map access and 8% in growing arrays.