@alex 2020-05-09
After this change, `MarshalCSV` is 4% of my runtime.