@alex 2020-05-10
Ran benchmarks from #257:

```
name         old time/op    new time/op    delta
WriteCSV-44    10.2ms ± 7%     8.0ms ± 5%  -22.06%  (p=0.000 n=19+20)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    10.7MB ± 0%     7.8MB ± 0%  -26.87%  (p=0.000 n=20+20)

name         old allocs/op  new allocs/op  delta
WriteCSV-44      103k ± 0%      102k ± 0%   -0.60%  (p=0.000 n=20+19)
```